requirements.txt
streamlit==1.35.0
streamlit-authenticator==0.3.2
pymongo[srv]==3.11
python-dotenv==1.0.1
watchdog==4.0.1
openai==1.34.0